+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [unreleased]
 
 - install missing environments before start build / test process
+- `bit build` with no parameter, builds all authored and imported components regardless whether they're modified
 
 ## [0.12.2-dev.1] - 2018-01-23